feat: add bowl form component
